==> 2_data_preparation.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
from ipywidgets import widgets, interactive
import missingno as msno
warnings.filterwarnings("ignore")
plt.style.use('fivethirtyeight')
pd.set_option('display.max_columns', None)
# pd.set_option('display.max_colwidth', None)

# Load dataset
train = pd.read_csv('train_attacker_2022_cleaned.csv')
test = pd.read_csv('test_attacker_2022_cleaned.csv')

print(train.shape)

# Split train data into train and validation sets
validation = train.sample(17459,random_state=44)
train = train.drop(validation.index)

print(train.shape)
print(validation.shape)

train_categorical = train.select_dtypes('object')
validation_categorical = validation.select_dtypes('object')
test_categorical = test.select_dtypes('object')

- Version 1: One hot encoding + PCA

# One hot encoding
from sklearn.preprocessing import OneHotEncoder

OHE = OneHotEncoder(drop='first',sparse=False,handle_unknown='ignore')
OHE.fit(train_categorical)

train_OHE = pd.DataFrame(OHE.transform(train_categorical),columns=OHE.get_feature_names_out(),index=train_categorical.index)
validation_OHE = pd.DataFrame(OHE.transform(validation_categorical),columns=OHE.get_feature_names_out(),index=validation_categorical.index)
test_OHE = pd.DataFrame(OHE.transform(test_categorical),columns=OHE.get_feature_names_out(),index=test_categorical.index)

train_OHE

# PCA
from sklearn.decomposition import PCA

pca = PCA(n_components=8)
pca.fit(train_OHE)

train_PCA = pd.DataFrame(pca.transform(train_OHE),index=train_OHE.index,columns=[f'pca_{i}' for i in range(1,9)])
validation_PCA = pd.DataFrame(pca.transform(validation_OHE),index=validation_OHE.index,columns=[f'pca_{i}' for i in range(1,9)])
test_PCA = pd.DataFrame(pca.transform(test_OHE),index=test_OHE.index,columns=[f'pca_{i}' for i in range(1,9)])

train_PCA

# Merge PCA variables to the raw data
train_v1 = pd.merge(train.select_dtypes(exclude='object'),train_PCA,left_index=True,right_index=True)
validation_v1 = pd.merge(validation.select_dtypes(exclude='object'),validation_PCA,left_index=True,right_index=True)
test_v1 = pd.merge(test.select_dtypes(exclude='object'),test_PCA,left_index=True,right_index=True)

train_v1

train_v1.to_csv('data_for_model/train_v1.csv',index=False)
validation_v1.to_csv('data_for_model/validation_v1.csv',index=False)
test_v1.to_csv('data_for_model/test_v1.csv',index=False)

- Version 2: Label encoding

train['location_id'].value_counts()

def label_encoding(feature):
    category_list = []
    data = train.copy()
    for i in data[feature].value_counts().index:
        category_list.append(data[data[feature] == i]['label'].value_counts(normalize=True).sort_index().to_list())
    df_encode = pd.DataFrame(category_list,index=data[feature].value_counts().index).fillna(0).sort_values(1,ascending=True)
    
    return list(df_encode.index)

from sklearn.preprocessing import OrdinalEncoder
for col in train_categorical.select_dtypes('object').columns:
    label_enc = OrdinalEncoder(categories=[label_encoding(col)],handle_unknown='use_encoded_value',unknown_value=-1)
    train_categorical[col] = label_enc.fit_transform(train_categorical[[col]]).flatten()
    validation_categorical[col] = label_enc.transform(validation_categorical[[col]]).flatten()
    test_categorical[col] = label_enc.transform(test_categorical[[col]]).flatten()

# Merge label encoding data to the raw data
train_v2 = pd.merge(train.select_dtypes(exclude='object'),train_categorical,left_index=True,right_index=True)
validation_v2 = pd.merge(validation.select_dtypes(exclude='object'),validation_categorical,left_index=True,right_index=True)
test_v2 = pd.merge(test.select_dtypes(exclude='object'),test_categorical,left_index=True,right_index=True)

train_v2

train_v2.to_csv('data_for_model/train_v2.csv',index=False)
validation_v2.to_csv('data_for_model/validation_v2.csv',index=False)
test_v2.to_csv('data_for_model/test_v2.csv',index=False)
